Fix incomplete list of placements being sent to the server
Looking at TrackerData, we were missing a call to AddTransitions.
This adds entries to the MainUpdater that manage state for
transitions; this may explain why a randomizer update (presumably
the state logic one?) broke our logic.

With the added call, every non-start placement in the spoiler log
is sent, and appears in the shuffled spoiler log produced by the
server.